+++ core/src/main/java/org/trade/core/persistent/dao/AccountRepository.java
@@ -2,7 +2,7 @@ package org.trade.core.persistent.dao;
 
 import org.trade.core.dao.AspectRepository;
 
-public interface AccountRepository extends AspectRepository<Account, Integer>, AccountRepositoryCustom {
+public interface AccountRepository extends AspectRepository<Account, Long>, AccountRepositoryCustom {
 
     Account findByAccountNumber(String accountNumber);
 }
