+++ core/src/main/java/org/trade/core/persistent/dao/StrategyRepository.java
@@ -3,7 +3,7 @@ package org.trade.core.persistent.dao;
 import org.trade.core.dao.AspectRepository;
 
 
-public interface StrategyRepository extends AspectRepository<Strategy, Integer>, StrategyRepositoryCustom {
+public interface StrategyRepository extends AspectRepository<Strategy, Long>, StrategyRepositoryCustom {
 
     Strategy findByName(String name);
 }
