+++ core/src/main/java/org/trade/core/persistent/dao/TradeOrderRepository.java
@@ -3,7 +3,7 @@ package org.trade.core.persistent.dao;
 import org.trade.core.dao.AspectRepository;
 
 
-public interface TradeOrderRepository extends AspectRepository<TradeOrder, Integer>, TradeOrderRepositoryCustom {
+public interface TradeOrderRepository extends AspectRepository<TradeOrder, Long>, TradeOrderRepositoryCustom {
 
     TradeOrder findByOrderKey(Integer orderKey);
 }
