+++ core/src/main/java/org/trade/core/persistent/dao/TradelogSummaryRepository.java
@@ -3,6 +3,6 @@ package org.trade.core.persistent.dao;
 import org.trade.core.dao.AspectRepository;
 
 
-public interface TradelogSummaryRepository extends AspectRepository<TradelogSummary, Integer>, TradelogSummaryRepositoryCustom {
+public interface TradelogSummaryRepository extends AspectRepository<TradelogSummary, Long>, TradelogSummaryRepositoryCustom {
 
 }
