+++ core/src/main/java/org/trade/core/persistent/dao/TradingdayRepository.java
@@ -3,6 +3,6 @@ package org.trade.core.persistent.dao;
 import org.trade.core.dao.AspectRepository;
 
 
-public interface TradingdayRepository extends AspectRepository<Tradingday, Integer>, TradingdayRepositoryCustom {
+public interface TradingdayRepository extends AspectRepository<Tradingday, Long>, TradingdayRepositoryCustom {
 
 }
